requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
ipython==8.21.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
matplotlib-inline==0.1.6
multidict==6.0.5
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
pytest==7.4.4
pytest-asyncio==0.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
pyzmq==25.1.2
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.36.3
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
wcwidth==0.2.13
yarl==1.9.4
